+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGES
 
+## 0.0.2
+
+- Change the `Contact` object to `ContactInfo`.
+- Change the `Item` object to `ValueItem`.
+
 ## 0.0.1
 
 - Release the package.